requirements.txt
appnope==0.1.2
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2
django-debug-toolbar==3.2
django-extensions==3.1.1
django-heroku==0.3.1
gunicorn==20.0.4
ipython==7.16.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1
traitlets==4.3.3
wcwidth==0.2.5
whitenoise==5.2.0
djangorestframework==3.12.2
pygments==2.8.1
httpie==2.4.0
graphene==2.1.8
graphene-django==2.15.0